feat: election results and types
