+++ frontend/src/Interfaces/Election.ts
@@ -1,39 +1,4 @@
-export default interface User {
-  id: number;
-  email: string;
-  dob: string;
-  staff: boolean;
-  superuser: boolean;
-}
-
-export interface Token {
-  access: string;
-  refresh: string;
-}
-
-export interface University {
-  id: number;
-  name: string;
-}
-
-export interface LoginParams {
-  email: string;
-  password: string;
-}
-
-export interface LoginResponse extends User {
-  token: Token;
-  university: University;
-}
-
-export interface createUser{
-  email: string,
-  password: string,
-  passwordConfirm: string,
-  universityId: number,
-  dob: Date,
-  admin: Boolean
-}
+import { University } from "./User";
 
 export interface Vote {
   electionId: number;
@@ -75,13 +40,6 @@ export interface Position {
   maxVotesPerCandidate: number;
 }
 
-export interface ElectionData {
-  code: number;
-  success: boolean;
-  msg: string;
-  data: ElectionDetail[];
-}
-
 export interface ElectionDetail {
   id: number;
   positions: PositionDetail[];
@@ -111,9 +69,7 @@ export interface CandidateDetail {
   candidateName: string;
   candidateDesc: string;
   photoUrl: string;
-}
-
-export interface University {
-  id: number;
-  name: string;
+  // election result
+  winner?: boolean;
+  votePercentage?: number;
 }
